user-stash-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "user_stash"
version = "0.0.1"
authors = [
  { name="Andre", email="xavier.andre256@gmail.com" },
]
maintainers = [
  { name="Andre", email="xavier.andre256@gmail.com" },
]
description = "User Stash Server"
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["flask"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
"flask==3.0.3",
"flask_cors==5.0.0",
"flask_sqlalchemy==3.1.1",
"flask_migrate",
"psycopg2-binary",
"pyjwt",
"python-dotenv"
]

[project.urls]
Repository = "https://github.com/AndreXVR/user-stash"
Issues = "https://github.com/AndreXVR/user-stash/issues"

[project.scripts]
user_stash_server = "user_stash.app:run"
